% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pixel_states.R
\name{compile_pixel_states}
\alias{compile_pixel_states}
\title{Deriving qualitative pixel states from raster data}
\usage{
compile_pixel_states(x, w = matrix(1, 3, 3), split_IQR = FALSE,
  custum_rclmat = NULL, sample = FALSE, size = 1000, ties = "NA",
  op = c("sampler", "proba"), inparallel = NULL, ...)
}
\arguments{
\item{x}{a raster stack or brick}

\item{w}{matrix of weights (the moving window), e.g. a 3 by 3 matrix with values 1; see Details. The matrix does not need to be square, but the sides must be odd numbers. If you need even sides, you can add a column or row with weights of zero}

\item{split_IQR}{logical. Should the inter-quartile range be split at the median
to form different ranges? The default is TRUE}

\item{custum_rclmat}{numeric vector indicating the limits based on which to reclassify x.}

\item{sample}{Logical. Should the data be sampled? If TRUE, the result is
based on a ramdon sampling of the original data.}

\item{size}{Integer. Sample size (if sample is TRUE or if the data cannot
be process in memory).}

\item{ties}{character. Indicates how to treat ties. Either 'random', 'lowest', 'highest', 'first', or 'NA'}

\item{op}{character string indicating the type of returned value: either raw states or their probabilities}

\item{inparallel}{integer indicating the number of processes to run in parallel}

\item{...}{Additional arguments as for writeRaster}
}
\value{
A raster brick where each layers correspond to one of the states.
}
\description{
make_pixel_states takes a raster stack or brick to qualify pixel
following scale either defined by the user or derived from the natural
breaks of data following the following boxplot statistics and outliers.
}
\details{
The ranges are bounded by the lower outlier (if any), extreme of the lower whisker,
the lower ‘hinge’, the median, the upper ‘hinge’, the extreme of the upper whisker,
and the upper outlier (if any).
}
\examples{
sample <- rnorm(100)
get_boxplot_range_1d (sample)
}
\author{
Issoufou Liman
}
