% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_raster.R
\name{discretize_raster}
\alias{discretize_raster}
\title{Discretize raster}
\usage{
discretize_raster(x, split_IQR = FALSE)
}
\arguments{
\item{x}{filename (character), Extent, Raster*, sf, SpatialPixels*, SpatialGrid*, object, 'image', matrix, im, or missing. Supported file types are the 'native' raster package format and those that can be read via \code{rgdal} (see \code{\link[rgdal]{readGDAL}}}

\item{split_IQR}{logical. Should the inter-quartile range be split at the median
to form different ranges? The default is TRUE}
}
\value{
A raster layer or list of raster layers
}
\description{
Discretize raster based boxplots statistics and outliers
}
\details{
The ranges are bounded by the lower outlier (if any), extreme of the lower whisker,
the lower ‘hinge’, the median, the upper ‘hinge’, the extreme of the upper whisker,
and the upper outlier (if any).
}
\author{
Issoufou Liman
}
